Add unit toggle button
The idea is to have a button that changes the temperature units.
At first I tried with two different buttons that would toggle the hidden
display style. It worked. But the code felt messy.

I then refactored to one button with a toggle class. Code looks messy still,
but far less than before. Great candidate for a more elegant refactor.

A key parameter was avoiding jQuery for this particular part. Probably
why the code looks a bit messy, but I'm happy avoiding the dependency.

Resources

- https://www.jamestease.co.uk/blether/add-remove-or-toggle-classes-using-vanilla-javascript
- http://youmightnotneedjquery.com/#get_html
- http://callmenick.com/post/add-remove-classes-with-javascript-property-classlist

- https://stackoverflow.com/questions/6348494/addeventlistener-vs-onclick
- https://www.simonewebdesign.it/onclick-vs-addeventlistener/
- https://www.w3schools.com/js/js_htmldom_eventlistener.asp
- https://developer.mozilla.org/en-US/docs/Web/Events

- https://stackoverflow.com/questions/4397821/hide-show-an-anchor-dynamically